requirements.txt
streamlit>=1.30.0
crewai>=0.28.0
pyyaml>=6.0.1
langchain-openai>=0.1.0